requirements.txt
# Core Dependencies
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.0.3

# API Clients
openai==1.30.1
pinecone-client==5.0.1
pymongo==4.6.1

# NLP & ML
numpy==1.26.4
scikit-learn==1.4.0
sentence-transformers==2.5.1
torch>=2.0.0
transformers==4.38.0

# Text Processing
nltk==3.9.1
tiktoken==0.6.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.28.1
aiohttp==3.10.5

# Utilities
pandas==2.2.3
tqdm==4.66.1
tenacity==9.1.2
asyncio-throttle==0.1.1

# Optional (for advanced features)
llama-index==0.13.4
wikipedia==1.4.0